Completely implement battleArena
At this point ArenaTOW basically encapsulates all of battleArena and is no
longer dependant on battleArena to function. It has been brought back to
the point of function that it was at months ago however is completely
self-dependant because all of battleArena has been added to it. Moving
forward the goal is to remove as many peices of battleArena as possible.